+++ src/mako/http/response/senders/File.php
@@ -12,6 +12,7 @@ use mako\file\FileSystem;
 use mako\http\exceptions\HttpException;
 use mako\http\Request;
 use mako\http\Response;
+use mako\http\response\Status;
 
 use function basename;
 use function connection_aborted;
@@ -254,7 +255,7 @@ class File implements ResponseSenderInterface
 			// Not an acceptable range so we'll just send an empty response
 			// along with a "requested range not satisfiable" status
 
-			$response->setStatus(416);
+			$response->setStatus(Status::RANGE_NOT_SATISFIABLE);
 
 			$response->sendHeaders();
 		}
@@ -274,7 +275,7 @@ class File implements ResponseSenderInterface
 				// Valid range so we'll need to tell the client which range we're sending
 				// and set the content-length header value to the length of the byte range
 
-				$response->setStatus(206);
+				$response->setStatus(Status::PARTIAL_CONTENT);
 
 				$response->headers->add('Content-Range', sprintf('bytes %s-%s/%s', $range['start'], $range['end'], $this->fileSize));