+++ tests/unit/error/handlers/web/ProductionHandlerTest.php
@@ -15,6 +15,7 @@ use mako\http\Request;
 use mako\http\request\Headers as RequestHeaders;
 use mako\http\Response;
 use mako\http\response\Headers as ResponseHeaders;
+use mako\http\response\Status;
 use mako\tests\TestCase;
 use mako\view\ViewFactory;
 use Mockery;
@@ -61,7 +62,7 @@ class ProductionHandlerTest extends TestCase
 
 		$response->shouldReceive('setBody')->once()->with('rendered')->andReturn($response);
 
-		$response->shouldReceive('setStatus')->once()->with(500)->andReturn($response);
+		$response->shouldReceive('setStatus')->once()->with(Status::INTERNAL_SERVER_ERROR)->andReturn($response);
 
 		$response->shouldReceive('send')->once();
 
@@ -140,7 +141,7 @@ class ProductionHandlerTest extends TestCase
 
 		$response->shouldReceive('setBody')->once()->with('rendered')->andReturn($response);
 
-		$response->shouldReceive('setStatus')->once()->with(500)->andReturn($response);
+		$response->shouldReceive('setStatus')->once()->with(Status::INTERNAL_SERVER_ERROR)->andReturn($response);
 
 		$response->shouldReceive('send')->once();
 
@@ -200,7 +201,7 @@ class ProductionHandlerTest extends TestCase
 
 		$response->shouldReceive('setBody')->once()->with('rendered')->andReturn($response);
 
-		$response->shouldReceive('setStatus')->once()->with(405)->andReturn($response);
+		$response->shouldReceive('setStatus')->once()->with(Status::METHOD_NOT_ALLOWED)->andReturn($response);
 
 		(function () use ($responseHeaders): void {
 			$this->headers = $responseHeaders;
@@ -270,7 +271,7 @@ class ProductionHandlerTest extends TestCase
 
 		$response->shouldReceive('setBody')->once()->with('An error has occurred while processing your request.')->andReturn($response);
 
-		$response->shouldReceive('setStatus')->once()->with(500)->andReturn($response);
+		$response->shouldReceive('setStatus')->once()->with(Status::INTERNAL_SERVER_ERROR)->andReturn($response);
 
 		$response->shouldReceive('send')->once();
 
@@ -317,7 +318,7 @@ class ProductionHandlerTest extends TestCase
 
 		$response->shouldReceive('setBody')->once()->with('An error has occurred while processing your request.')->andReturn($response);
 
-		$response->shouldReceive('setStatus')->once()->with(500)->andReturn($response);
+		$response->shouldReceive('setStatus')->once()->with(Status::INTERNAL_SERVER_ERROR)->andReturn($response);
 
 		$response->shouldReceive('send')->once();
 
@@ -360,7 +361,7 @@ class ProductionHandlerTest extends TestCase
 
 		$response->shouldReceive('setBody')->once()->with('{"error":{"code":500,"message":"An error has occurred while processing your request.","exception_id":"exception_id"}}')->andReturn($response);
 
-		$response->shouldReceive('setStatus')->once()->with(500)->andReturn($response);
+		$response->shouldReceive('setStatus')->once()->with(Status::INTERNAL_SERVER_ERROR)->andReturn($response);
 
 		$response->shouldReceive('send')->once();
 
@@ -409,7 +410,7 @@ class ProductionHandlerTest extends TestCase
 
 		$response->shouldReceive('setBody')->once()->with('{"error":{"code":403,"message":"You don\'t have permission to access the requested resource.","exception_id":"exception_id","metadata":{"foo":"bar"}}}')->andReturn($response);
 
-		$response->shouldReceive('setStatus')->once()->with(403)->andReturn($response);
+		$response->shouldReceive('setStatus')->once()->with(Status::FORBIDDEN)->andReturn($response);
 
 		$response->shouldReceive('send')->once();
 
@@ -471,7 +472,7 @@ class ProductionHandlerTest extends TestCase
 <error><code>500</code><message>An error has occurred while processing your request.</message><exception_id>exception_id</exception_id></error>
 ')->andReturn($response);
 
-		$response->shouldReceive('setStatus')->once()->with(500)->andReturn($response);
+		$response->shouldReceive('setStatus')->once()->with(Status::INTERNAL_SERVER_ERROR)->andReturn($response);
 
 		$response->shouldReceive('send')->once();
 
@@ -533,7 +534,7 @@ class ProductionHandlerTest extends TestCase
 <error><code>403</code><message>You don\'t have permission to access the requested resource.</message><exception_id>exception_id</exception_id><metadata><foo>bar</foo></metadata></error>
 ')->andReturn($response);
 
-		$response->shouldReceive('setStatus')->once()->with(403)->andReturn($response);
+		$response->shouldReceive('setStatus')->once()->with(Status::FORBIDDEN)->andReturn($response);
 
 		$response->shouldReceive('send')->once();